# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Catalog": {
        "additionalProperties": true,
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionSpec"
            },
            "title": "Actions",
            "type": "array"
          },
          "capabilities": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Capabilities",
            "type": "object"
          },
          "dashboards": {
            "items": {
              "$ref": "#/components/schemas/Dashboard"
            },
            "title": "Dashboards",
            "type": "array"
          },
          "defaults": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "title": "Defaults",
            "type": "object"
          },
          "extension_sources": {
            "items": {
              "$ref": "#/components/schemas/Dashboard"
            },
            "title": "Extension Sources",
            "type": "array"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/FeatureSpec"
            },
            "title": "Features",
            "type": "array"
          },
          "icons": {
            "items": {
              "type": "string"
            },
            "title": "Icons",
            "type": "array"
          },
          "keys": {
            "$ref": "#/components/schemas/Keys"
          },
          "layouts": {
            "items": {
              "type": "string"
            },
            "title": "Layouts",
            "type": "array"
          },
          "limits": {
            "$ref": "#/components/schemas/Limits"
          },
          "locales": {
            "items": {
              "type": "string"
            },
            "title": "Locales",
            "type": "array"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "actions",
          "features",
          "icons",
          "keys",
          "dashboards",
          "extension_sources",
          "locales",
          "layouts",
          "sources",
          "capabilities",
          "limits",
          "defaults"
        ],
        "title": "Catalog",
        "type": "object"
      },
...
      "ConfigDocument": {
        "additionalProperties": true,
        "properties": {
          "features": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Features",
            "type": "object"
          },
          "integrations": {
            "$ref": "#/components/schemas/Integrations"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/PageDocument"
            },
            "title": "Pages",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "scripts": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Scripts",
            "type": "object"
          },
          "server": {
            "$ref": "#/components/schemas/ServerSettings"
          }
        },
        "required": [
          "revision",
          "server",
          "features",
          "integrations",
          "pages"
        ],
        "title": "ConfigDocument",
        "type": "object"
      },
      "ConfigInput": {
        "properties": {
          "features": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Features",
            "type": "object"
          },
          "integrations": {
            "$ref": "#/components/schemas/IntegrationsInput"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/PageInput"
            },
            "title": "Pages",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "scripts": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "readOnly": true,
            "title": "Scripts",
            "type": "object"
          },
          "server": {
            "$ref": "#/components/schemas/ServerInput"
          }
        },
        "required": [
          "pages"
        ],
        "title": "ConfigInput",
        "type": "object"
      },
      "ConfigRead": {
        "additionalProperties": true,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ConfigDocument"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "config",
          "path"
        ],
        "title": "ConfigRead",
        "type": "object"
      },
      "ConfigSaved": {
        "additionalProperties": true,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ConfigDocument"
          },
          "saved": {
            "title": "Saved",
            "type": "boolean"
          }
        },
        "required": [
          "saved",
          "config"
        ],
        "title": "ConfigSaved",
        "type": "object"
      },
      "Configure": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operation": {
            "const": "configure",
            "title": "Operation",
            "type": "string"
          },
          "settings": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Settings",
            "type": "object"
          }
        },
        "required": [
          "operation",
          "id",
          "settings"
        ],
        "title": "Configure",
        "type": "object"
      },
...
      "Enable": {
        "additionalProperties": false,
        "properties": {
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operation": {
            "const": "enable",
            "title": "Operation",
            "type": "string"
          },
          "trust": {
            "title": "Trust",
            "type": "boolean"
          }
        },
        "required": [
          "operation",
          "id",
          "trust",
          "digest"
        ],
        "title": "Enable",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "code",
          "request_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExtensionCatalog": {
        "additionalProperties": true,
        "properties": {
          "api_version": {
            "title": "Api Version",
            "type": "integer"
          },
          "directory": {
            "title": "Directory",
            "type": "string"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "title": "Errors",
            "type": "array"
          },
          "extensions": {
            "items": {
              "$ref": "#/components/schemas/InstalledExtension"
            },
            "title": "Extensions",
            "type": "array"
          }
        },
        "required": [
          "api_version",
          "directory",
          "extensions",
          "errors"
        ],
        "title": "ExtensionCatalog",
        "type": "object"
      },
...
      "Install": {
        "additionalProperties": false,
        "properties": {
          "operation": {
            "enum": [
              "install",
              "rescan"
            ],
            "title": "Operation",
            "type": "string"
          }
        },
        "required": [
          "operation"
        ],
        "title": "Install",
        "type": "object"
      },
...
      "IntegrationsInput": {
        "properties": {
          "obs": {
            "$ref": "#/components/schemas/ObsInput"
          }
        },
        "title": "IntegrationsInput",
        "type": "object"
      },
      "JsonBody": {
        "$ref": "#/components/schemas/JsonValue",
        "description": "Keep JSON recursive references scoped to a Pydantic model.",
        "title": "JsonBody"
      },
      "JsonValue": {},
...
      "OperationResult": {
        "additionalProperties": true,
        "properties": {
          "cancelled": {
            "default": false,
            "title": "Cancelled",
            "type": "boolean"
          },
          "ok": {
            "default": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "OperationResult",
        "type": "object"
      },
...
      "PageInput": {
        "properties": {
          "buttons": {
            "items": {
              "$ref": "#/components/schemas/ButtonInput"
            },
            "title": "Buttons",
            "type": "array"
          },
          "dashboard": {
            "title": "Dashboard",
            "type": "string"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "layout": {
            "enum": [
              "grid",
              "list"
            ],
            "title": "Layout",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": {
                  "type": "string"
                },
                "propertyNames": {
                  "enum": [
                    "en",
                    "fr"
                  ]
                },
                "type": "object"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id"
        ],
        "title": "PageInput",
        "type": "object"
      },
...
      "Remove": {
        "additionalProperties": false,
        "properties": {
          "confirm": {
            "title": "Confirm",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operation": {
            "const": "remove",
            "title": "Operation",
            "type": "string"
          }
        },
        "required": [
          "operation",
          "id",
          "confirm"
        ],
        "title": "Remove",
        "type": "object"
      },
...
      "ServerInput": {
        "properties": {
          "host": {
            "title": "Host",
            "type": "string"
          },
          "poll_interval": {
            "title": "Poll Interval",
            "type": "number"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "volume_step": {
            "title": "Volume Step",
            "type": "integer"
          }
        },
        "title": "ServerInput",
        "type": "object"
      },
...
      "TargetOperation": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operation": {
            "enum": [
              "disable",
              "restart"
            ],
            "title": "Operation",
            "type": "string"
          }
        },
        "required": [
          "operation",
          "id"
        ],
        "title": "TargetOperation",
        "type": "object"
      },
...
      "ValidationResult": {
        "additionalProperties": true,
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          }
        },
        "required": [
          "valid",
          "error"
        ],
        "title": "ValidationResult",
        "type": "object"
      }
    },
    "securitySchemes": {
      "LocalSession": {
        "description": "Per-start local UI session; never the durable console pairing token.",
        "in": "header",
        "name": "X-Deck3DS-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Authenticated loopback API. Console TCP and extension stdio protocols remain independent.",
    "title": "3Decks local agent API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/config": {
      "get": {
        "operationId": "get_config",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Read",
        "tags": [
          "configuration"
        ]
      },
      "put": {
        "operationId": "save_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigSaved"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Save",
        "tags": [
          "configuration"
        ]
      }
    },
    "/api/config/validate": {
      "post": {
        "operationId": "validate_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Validate",
        "tags": [
          "configuration"
        ]
      }
    },
    "/api/extensions": {
      "get": {
        "operationId": "get_extensions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionCatalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Extensions",
        "tags": [
          "extensions"
        ]
      },
      "post": {
        "operationId": "manage_extension",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "configure": "#/components/schemas/Configure",
                    "disable": "#/components/schemas/TargetOperation",
                    "enable": "#/components/schemas/Enable",
                    "install": "#/components/schemas/Install",
                    "remove": "#/components/schemas/Remove",
                    "rescan": "#/components/schemas/Install",
                    "restart": "#/components/schemas/TargetOperation"
                  },
                  "propertyName": "operation"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Install"
                  },
                  {
                    "$ref": "#/components/schemas/TargetOperation"
                  },
                  {
                    "$ref": "#/components/schemas/Enable"
                  },
                  {
                    "$ref": "#/components/schemas/Configure"
                  },
                  {
                    "$ref": "#/components/schemas/Remove"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Manage",
        "tags": [
          "extensions"
        ]
      }
    },
...
    "/api/schema": {
      "get": {
        "operationId": "get_catalog",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Malformed body"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Local session refused"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Method not allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict or unavailable integration"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request too large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation failed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Agent unavailable"
          }
        },
        "summary": "Schema",
        "tags": [
          "configuration"
        ]
      }
    },
...
  },
  "security": [
    {
      "LocalSession": []
    }
  ]
}